refactor: Load More func on Category and Offers pg
Both these pages fetch for listings as soon as component mounts
creating a reference to 'listings' collection, to create a query
to the Firebase db to pull the listings either by 'type' or where
'offer' is true.

Listings are fetched with a limit which can be adjusted, if
listings still remain, user can 'load more' and the screen will scroll
to the bottom to continue fetching listings to display.

Also added the listings count, so the user can see how many listings
are already available to scroll through and view.